Rubocop clustering controller

+++ app/controllers/clustering_controller.rb
@@ -1,8 +1,8 @@
 require 'clustering/clustering'
 
+# This is the controller handles the automatic clustring BL
 class ClusteringController < ApplicationController
-  
-  authorize_resource :class => false
+  authorize_resource class: false
 
   def select
   end
@@ -10,46 +10,49 @@ class ClusteringController < ApplicationController
   def confirm
     @clusters = Clustering.run_algorithm params[:algorithm], params[:kappa]
   end
-  
+
   def edit
     @clusters = Cluster.all
   end
 
   def save
-    begin
-      ActiveRecord::Base.transaction do
-        params[:clusterprosumers].zip(params[:clusternames], params[:clusterdescriptions], params[:clusterids]).each do |prosumers, clustername, desc, clid|
-          cluster = clid.to_i > 0 ? Cluster.find(clid) : Cluster.new
-          cluster.name = clustername
-          cluster.description = desc
-          prs = Prosumer.find( prosumers.split(",") )
-          cluster.prosumers << prs
-          cluster.save!
-        end
-        
-        Cluster.all.each do |cl|
-          if cl.prosumers.size == 0
-            cl.destroy
-          end
-        end
-        
-        respond_to do |format|
-          format.html { redirect_to clusters_path, notice: 'Clusters were successfully updated.' }
+    ActiveRecord::Base.transaction do
+      params[:clusterprosumers].zip(
+        params[:clusternames], params[:clusterdescriptions],
+        params[:clusterids]).each do |prosumers, clustername, desc, clid|
+        cluster = clid.to_i > 0 ? Cluster.find(clid) : Cluster.new
+        cluster.name = clustername
+        cluster.description = desc
+        prs = Prosumer.find(prosumers.split(','))
+        cluster.prosumers << prs
+        cluster.save!
+      end
+
+      Cluster.all.each do |cl|
+        cl.destroy if cl.prosumers.size == 0
+      end
+
+      respond_to do |format|
+        format.html do
+          redirect_to clusters_path,
+                      notice: 'Clusters were successfully updated.'
         end
       end
-    rescue 
-      respond_to do |format|
-        format.html { redirect_to "/clustering/select", alert: 'Clusters were NOT successfully updated.' }
+    end
+  rescue
+    respond_to do |format|
+      format.html do
+        redirect_to '/clustering/select',
+                    alert: 'Clusters were NOT successfully updated.'
       end
     end
   end
 
+  helper_method :algorithms
 
-  helper_method :algorithms 
-  
   private
-    def algorithms
-      Clustering.algorithms
-    end
- 
+
+  def algorithms
+    Clustering.algorithms
+  end
 end
